move any grammar or syntax related code to healther
