+++ python/.pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
     rev: v1.18.1
     hooks:
       - id: mypy
-        files: ^python/src/hhat_lang/toolchain/
+        files: ^python/src/hhat_lang/(toolchain|dialects/heather/grammar/doc_signatures\.py)
         args: [
           --ignore-missing-imports,
           --disable-error-code, attr-defined,
@@ -22,10 +22,10 @@ repos:
     rev: "v0.14.9"
     hooks:
       - id: ruff-check
-        files: ^python/(src/hhat_lang/toolchain|tests/test_cli\.py$)
+        files: ^python/(src/hhat_lang/(toolchain|dialects/heather/grammar/doc_signatures\.py)|tests/test_cli\.py$)
         args: [--fix, --show-fixes, --show-files]
       - id: ruff-format
-        files: ^python/(src/hhat_lang/toolchain|tests/test_cli\.py$)
+        files: ^python/(src/hhat_lang/(toolchain|dialects/heather/grammar/doc_signatures\.py)|tests/test_cli\.py$)
         args: [--line-length=100]
 
 default_language_version:
